location: extract backend specific part of StripPassword
The tests for the rest backend now reside there.
